Don't use broken zoomFov playerState field; smooth disruptor zoom-out
It was used as float, but sent as 8bit unsigned integer over the
net. In calculations it's value was always a positive epsilon.

+++ code/game/q_shared.h
@@ -1686,7 +1686,7 @@ typedef struct playerState_s {
 	int			zoomMode;		// 0 - not zoomed, 1 - disruptor weapon
 	int			zoomTime;
 	qboolean	zoomLocked;
-	float		zoomFov;
+	int			zoomFov;		// unused, was float
 	int			zoomLockTime;
 
 	int			fallingToDeath;
